Move ConnectorException into Steeltoe.Connectors.CloudFoundry
